requirements.txt
python-socketio>=5.0
eventlet>=0.33
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
